+++ src/main/kotlin/things/Cells/State.kt
@@ -1,5 +1,5 @@
 import things.Microorganisms
 
-data class State(var x: Int, var y: Int, var focused: Microorganisms?){
+data class State(var x: Double, var y: Double, var focused: Microorganisms?){
 
 }
